Add lots of window functions, and move all window functions into its own file
